+++ src/Commands/CommandRegistration.cs
@@ -1,8 +1,11 @@
 ﻿using System.ComponentModel.Composition;
+using System.Windows.Threading;
 using Microsoft.VisualStudio.Editor;
+using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Text;
 using Microsoft.VisualStudio.Text.Classification;
 using Microsoft.VisualStudio.Text.Editor;
+using Microsoft.VisualStudio.Text.Operations;
 using Microsoft.VisualStudio.TextManager.Interop;
 using Microsoft.VisualStudio.Utilities;
 
@@ -14,7 +17,7 @@ namespace MarkdownEditor
     public class CommandRegistration : IVsTextViewCreationListener
     {
         [Import]
-        public IVsEditorAdaptersFactoryService EditorAdaptersFactoryService { get; set; }
+        IVsEditorAdaptersFactoryService EditorAdaptersFactoryService { get; set; }
 
         [Import]
         ITextDocumentFactoryService TextDocumentFactoryService { get; set; }
@@ -22,21 +25,38 @@ namespace MarkdownEditor
         [Import]
         IClassifierAggregatorService ClassifierAggregatorService { get; set; }
 
+        [Import]
+        ITextStructureNavigatorSelectorService NavigatorService { get; set; }
+
         public void VsTextViewCreated(IVsTextView textViewAdapter)
         {
-            var textView = EditorAdaptersFactoryService.GetWpfTextView(textViewAdapter);
-            ITextDocument document;
-
-            if (TextDocumentFactoryService.TryGetTextDocument(textView.TextBuffer, out document))
+            ThreadHelper.Generic.BeginInvoke(DispatcherPriority.ApplicationIdle, () =>
             {
+                var textView = EditorAdaptersFactoryService.GetWpfTextView(textViewAdapter);
+                ITextDocument document = null;
+
+                if (textView == null || !TextDocumentFactoryService.TryGetTextDocument(textView.TextBuffer, out document))
+                    return;
+
                 textView.Properties.GetOrCreateSingletonProperty(() => new PasteImage(textViewAdapter, textView, document.FilePath));
-                textView.Properties.GetOrCreateSingletonProperty(() => new BoldCommandTarget(textViewAdapter, textView));
-                textView.Properties.GetOrCreateSingletonProperty(() => new ItalicCommandTarget(textViewAdapter, textView));
-                textView.Properties.GetOrCreateSingletonProperty(() => new InlineCodeCommandTarget(textViewAdapter, textView));
+                textView.Properties.GetOrCreateSingletonProperty(() => new BoldCommandTarget(textViewAdapter, textView, NavigatorService));
+                textView.Properties.GetOrCreateSingletonProperty(() => new ItalicCommandTarget(textViewAdapter, textView, NavigatorService));
+                textView.Properties.GetOrCreateSingletonProperty(() => new InlineCodeCommandTarget(textViewAdapter, textView, NavigatorService));
                 textView.Properties.GetOrCreateSingletonProperty(() => new SmartIndentCommandTarget(textViewAdapter, textView));
                 textView.Properties.GetOrCreateSingletonProperty(() => new IndentationCommandTarget(textViewAdapter, textView));
                 textView.Properties.GetOrCreateSingletonProperty(() => new ToogleTaskCommandTarget(textViewAdapter, textView));
                 textView.Properties.GetOrCreateSingletonProperty(() => new Navigate(textViewAdapter, textView));
+
+                document.FileActionOccurred += Document_FileActionOccurred;
+            });
+        }
+
+        private void Document_FileActionOccurred(object sender, TextDocumentFileActionEventArgs e)
+        {
+            if (e.FileActionType == FileActionTypes.ContentSavedToDisk)
+            {
+                if (GenerateHtml.HtmlGenerationEnabled(e.FilePath))
+                    GenerateHtml.GenerateHtmlFile(e.FilePath);
             }
         }
     }