+++ src/Commands/ItalicCommandTarget.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.Text.Editor;
+using Microsoft.VisualStudio.Text.Operations;
 using Microsoft.VisualStudio.TextManager.Interop;
 
 namespace MarkdownEditor
@@ -10,13 +11,13 @@ namespace MarkdownEditor
         private static readonly Guid _commandGroup = typeof(VSConstants.VSStd2KCmdID).GUID;
         private static readonly uint _commandId = (uint)VSConstants.VSStd2KCmdID.ISEARCH; // maps to Edit.IncrementalSearch
 
-        public ItalicCommandTarget(IVsTextView adapter, IWpfTextView textView)
-            :base (adapter, textView, _commandGroup, _commandId)
+        public ItalicCommandTarget(IVsTextView adapter, IWpfTextView textView, ITextStructureNavigatorSelectorService navigator)
+            :base (adapter, textView, navigator, _commandGroup, _commandId)
         { }
 
         public override string Symbol
         {
-            get { return "_"; }
+            get { return MarkdownEditorPackage.Options.BoldStyle == EmphasisStyle.Asterisk ? "*" : "_"; }
         }
     }
 }