+++ src/LanguageService/ContentTypeDefinition.cs
@@ -5,6 +5,8 @@ namespace MarkdownEditor
 {
     public class ContentTypeDefinition
     {
+        public static string[] MarkdownExtensions = { ".markdown", ".md", ".mdown", ".mdwn", ".mkd", ".mkdn", ".mmd", ".rst" };
+
         [Export(typeof(ContentTypeDefinition))]
         [Name(MarkdownLanguage.LanguageName)]
         [BaseDefinition("text")]