+++ src/MarkdownEditorPackage.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.ComponentModel.Design;
 using System.Runtime.InteropServices;
-using System.Windows.Threading;
 using Microsoft.VisualStudio;
 using Microsoft.VisualStudio.Shell;
 using Microsoft.VisualStudio.Shell.Interop;
@@ -9,9 +8,10 @@ using Microsoft.VisualStudio.Shell.Interop;
 namespace MarkdownEditor
 {
     [Guid(PackageGuids.guidPackageString)]
-    [PackageRegistration(UseManagedResourcesOnly = true)]
+    [PackageRegistration(UseManagedResourcesOnly = true, AllowsBackgroundLoading = true)]
     [InstalledProductRegistration("#110", "#112", Vsix.Version, IconResourceID = 400)]
-    [ProvideService(typeof(MarkdownLanguage), ServiceName = MarkdownLanguage.LanguageName)]
+    [ProvideMenuResource("Menus.ctmenu", 1)]
+
     [ProvideLanguageService(typeof(MarkdownLanguage), MarkdownLanguage.LanguageName, 100, ShowDropDownOptions = true, DefaultToInsertSpaces = true, EnableCommenting = true, AutoOutlining = true)]
     [ProvideLanguageEditorOptionPage(typeof(Options), MarkdownLanguage.LanguageName, null, "Advanced", "#101", new[] { "markdown", "md" })]
     [ProvideLanguageExtension(typeof(MarkdownLanguage), ".markdown")]
@@ -21,7 +21,35 @@ namespace MarkdownEditor
     [ProvideLanguageExtension(typeof(MarkdownLanguage), ".mkd")]
     [ProvideLanguageExtension(typeof(MarkdownLanguage), ".mkdn")]
     [ProvideLanguageExtension(typeof(MarkdownLanguage), ".mmd")]
-    [ProvideMenuResource("Menus.ctmenu", 1)]
+    [ProvideLanguageExtension(typeof(MarkdownLanguage), ".rst")]
+
+    [ProvideEditorFactory(typeof(EditorFactory), 110, CommonPhysicalViewAttributes = (int)__VSPHYSICALVIEWATTRIBUTES.PVA_None, TrustLevel = __VSEDITORTRUSTLEVEL.ETL_AlwaysTrusted)]
+    [ProvideEditorLogicalView(typeof(EditorFactory), VSConstants.LOGVIEWID.TextView_string, IsTrusted = true)]
+
+    [ProvideEditorExtension(typeof(EditorFactory), ".markdown", 100)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".md", 100)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".mdown", 100)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".mdwn", 100)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".mkd", 100)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".mkdn", 100)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".mmd", 100)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".rst", 100)]
+    [ProvideEditorExtension(typeof(EditorFactory), ".*", 2, NameResourceID = 110)]
+
+    [ProvideAutoLoad("559ffa40-ab05-4ca2-9cc6-fb20c4a37112")]
+    [ProvideUIContextRule("559ffa40-ab05-4ca2-9cc6-fb20c4a37112",
+    name: Vsix.Name,
+    expression: "(markdown | md | mdown | mdwn | mkd | mmd | rst)",
+    termNames: new[] { "markdown", "md", "mdown", "mdwn", "mkd", "mkdn", "mmd", "rst" },
+    termValues: new[] {
+        "HierSingleSelectionName:.markdown$",
+        "HierSingleSelectionName:.md$",
+        "HierSingleSelectionName:.mdown$",
+        "HierSingleSelectionName:.mdwn$",
+        "HierSingleSelectionName:.mkd$",
+        "HierSingleSelectionName:.mkdn$",
+        "HierSingleSelectionName:.mmd$",
+        "HierSingleSelectionName:.rst$"})]
     public sealed class MarkdownEditorPackage : Package
     {
         private static Options _options;
@@ -37,10 +65,7 @@ namespace MarkdownEditor
                     {
                         if (_options == null)
                         {
-                            // Load the package when options are needed
-                            var shell = (IVsShell)GetGlobalService(typeof(SVsShell));
-                            IVsPackage package;
-                            ErrorHandler.ThrowOnFailure(shell.LoadPackage(ref PackageGuids.guidPackage, out package));
+                            LoadPackage();
                         }
                     }
                 }
@@ -54,15 +79,27 @@ namespace MarkdownEditor
             _options = (Options)GetDialogPage(typeof(Options));
 
             Logger.Initialize(this, Vsix.Name);
+            ErrorList.Initialize(this);
             CopyAsHtmlCommand.Initialize(this);
             AddCustomStylesheet.Initialize(this);
+            GenerateHtml.Initialize(this);
 
             var serviceContainer = this as IServiceContainer;
-            var langService = new MarkdownLanguage();
-            langService.SetSite(this);
+            var langService = new MarkdownLanguage(this);
             serviceContainer.AddService(typeof(MarkdownLanguage), langService, true);
 
-            base.Initialize();
+            var editorFactory = new EditorFactory(this, typeof(MarkdownLanguage).GUID);
+            RegisterEditorFactory(editorFactory);
+        }
+
+        private static void LoadPackage()
+        {
+            var shell = (IVsShell)GetGlobalService(typeof(SVsShell));
+
+            IVsPackage package;
+
+            if (shell.IsPackageLoaded(ref PackageGuids.guidPackage, out package) != VSConstants.S_OK)
+                ErrorHandler.Succeeded(shell.LoadPackage(ref PackageGuids.guidPackage, out package));
         }
     }
 }